Use UInt128 for BytesProcessed and set default MaxBytesLimit to ulong.MaxValue * 64

+++ src/CryptoBase/SymmetricCryptos/StreamCryptos/ChaCha20Crypto.cs
@@ -12,7 +12,7 @@ public class ChaCha20Crypto : SnuffleCrypto
 	/// <summary>
 	/// ChaCha20 uses a 32-bit counter, max blocks = 2^32, max bytes = 2^32 * 64
 	/// </summary>
-	protected override ulong MaxBytesLimit => (1UL << 32) * BlockSize;
+	protected override UInt128 MaxBytesLimit => (UInt128)(1UL << 32) * BlockSize;
 
 	public ChaCha20Crypto(ReadOnlySpan<byte> key, ReadOnlySpan<byte> iv)
 	{
@@ -118,7 +118,7 @@ public class ChaCha20Crypto : SnuffleCrypto
 
 	public void SetCounter(uint counter)
 	{
-		BytesProcessed = (ulong)counter * BlockSize;
+		BytesProcessed = (UInt128)counter * BlockSize;
 		Index = 0;
 		ChaCha20Utils.GetCounter(ref State.GetReference()) = counter;
 	}
